Readme updates
Having discussed this, it's better to assume the default/interactive
use case for the auth token (do not pass it on the command line and
save to history). Remove the argument from the recommendation.

Adding a note in case a context was used auth switch is then required
prior to provisioning.

Whitespace cleanups