Add comment about ansible-vault, and update the readme link
Readme link was a few lines too early (probably an insertion further
up), change line 64 to line 67 so the link points to the right
location.

I ran into a confusing situation running what looked like normal shell
commands to pass random data to ansible-vault, where a password was
prompted for, and decrypting the vault secrets in the yaml file
failed.

Add a safeguard note that you should be in the selfhost
directory (where the ansible.cfg file is present) before running the
shell commands, and a hint that being prompted for a password is an
indication that something went wrong.